tooling: expose a utility types module
We need these types in a few places, and we'd otherwise have a circular
dependency between `tooling/hcpctl` and `test`.
