+++ go.work
@@ -24,5 +24,6 @@ use (
 	./tooling/secret-sync
 	./tooling/templatize
 	./tooling/tenant-quota
+	./tooling/utilitytypes
 	./tooling/yamlwrap
 )
