+++ tooling/utilitytypes/go.mod
@@ -1,0 +1,3 @@
+module github.com/Azure/ARO-HCP/tooling/utilitytypes
+
+go 1.25.7
